Rename PeerVerifier to PeerVerifierOld
